add classchecker

+++ 275-jvm-test/src/main/java/com/njuse/jvmfinal/instructions/conversion/I2B.java
@@ -11,7 +11,7 @@ public class I2B extends NoOperandsInstruction {
      */
     @Override
     public void execute(StackFrame frame) {
-        short b=(short)(frame.getOperandStack().popInt()&0x000000FF);
+        byte b=(byte)(frame.getOperandStack().popInt()&0x000000FF);
         frame.getOperandStack().pushInt(b);
     }
 }
